1.65.4.1 チャンネル動画は投稿者名の代わりにチャンネル名称を使う。

+++ frontend/src/saccubus/ConvertWorker.java
@@ -1025,14 +1025,16 @@ public class ConvertWorker extends SwingWorker<String, String> {
 			return false;
 		}
 		System.out.println("reading:" + thumbInfo);
+		boolean isOK = true;
 		if(!saveThumbUser(thumbInfo, client)){
 			System.out.println("投稿者情報の取得に失敗");
-			return false;
+			isOK = false;
 		}
 		if(!saveThumbnailJpg(thumbInfo, client)){
 			System.out.println("サムネイル画像の取得に失敗");
-			return false;
+			isOK = false;
 		}
+		if(!isOK) return false;
 		//Path.fileCopy(thumbInfo, thumbInfoFile);
 		String text = Path.readAllText(thumbInfo.getPath(), "UTF-8");
 		text = text.replace("\n", "\r\n");
@@ -1044,7 +1046,9 @@ public class ConvertWorker extends SwingWorker<String, String> {
 			pw.close();
 		} catch (IOException e) {
 			e.printStackTrace();
+			isOK = false;
 		}
+		if(!isOK) return false;
 		if(thumbInfo.delete()){
 			System.out.println("Deleted:" + thumbInfo);
 		}
@@ -1068,9 +1072,19 @@ public class ConvertWorker extends SwingWorker<String, String> {
 	private boolean saveThumbUser(Path infoFile, NicoClient client) {
 		sendtext("投稿者情報の保存");
 		Path userThumbFile = null;
+		boolean isUser = true;
+		String ownerName = null;
 		if(Setting.isSaveThumbUser()){
 			String infoXml = Path.readAllText(infoFile.getPath(), "UTF-8");
 			String userID = NicoClient.getXmlElement(infoXml, "user_id");
+			String user_nickname = NicoClient.getXmlElement(infoXml, "user_nickname");
+			if(userID==null || userID.isEmpty()){
+				isUser = false;
+				userID = NicoClient.getXmlElement(infoXml, "ch_id");
+				ownerName = NicoClient.getXmlElement(infoXml, "ch_name");
+				if(userID!=null && !userID.isEmpty())
+					userID = "ch"+userID;
+			}
 			if(userID==null || userID.isEmpty() || userID.equals("none")){
 				sendtext("投稿者の情報がありません");
 				result = "A5";
@@ -1088,37 +1102,43 @@ public class ConvertWorker extends SwingWorker<String, String> {
 			}
 			userThumbFile = new Path(userFolder, userID + ".htm");
 			String html = null;
-			String ownerName = null;
-			if(!userThumbFile.canRead()){
-				userThumbFile = client.getThumbUserFile(userID, userFolder);
-			}
-			if(userThumbFile != null && userThumbFile.canRead()){
-				html = Path.readAllText(userThumbFile.getPath(), "UTF-8");
-				ownerName = NicoClient.getXmlElement(html, "title");
-			}
-			if(ownerName == null || ownerName.contains("非公開プロフィール")){
-				ownerName = null;
-				userThumbFile = client.getUserInfoFile(userID, userFolder);
+			if(isUser){
+				if(!userThumbFile.canRead()){
+					userThumbFile = client.getThumbUserFile(userID, userFolder);
+				}
 				if(userThumbFile != null && userThumbFile.canRead()){
 					html = Path.readAllText(userThumbFile.getPath(), "UTF-8");
 					ownerName = NicoClient.getXmlElement(html, "title");
 				}
-				if(ownerName==null){
-					sendtext("投稿者の情報の入手に失敗");
-					result = "A7";
-					return false;
+				if(ownerName == null || ownerName.contains("非公開プロフィール")){
+					ownerName = null;
+					userThumbFile = client.getUserInfoFile(userID, userFolder);
+					if(userThumbFile != null && userThumbFile.canRead()){
+						html = Path.readAllText(userThumbFile.getPath(), "UTF-8");
+						ownerName = NicoClient.getXmlElement(html, "title");
+					}
+					if(ownerName==null){
+						sendtext("投稿者の情報の入手に失敗");
+						result = "A7";
+						if(ownerName==null || ownerName.isEmpty())
+							ownerName = user_nickname;
+						if(ownerName==null || ownerName.isEmpty())
+							ownerName = "投稿者の情報の入手に失敗";
+					//	return false;
+					}
 				}
+				int index = ownerName.lastIndexOf("さんのプロフィール‐");
+				if(index > 0){
+					ownerName = ownerName.substring(0,index);
+				}
+				index = ownerName.lastIndexOf("さんのユーザーページ ‐");
+				if(index > 0){
+					ownerName = ownerName.substring(0,index) + "(ニコレポ非公開)";
+				}
+				if(user_nickname==null || user_nickname.isEmpty())
+					infoXml = infoXml.replace("</user_id>",
+						"</user_id>\n<user>" + ownerName + "</user>");
 			}
-			int index = ownerName.lastIndexOf("さんのプロフィール‐");
-			if(index > 0){
-				ownerName = ownerName.substring(0,index);
-			}
-			index = ownerName.lastIndexOf("さんのユーザーページ ‐");
-			if(index > 0){
-				ownerName = ownerName.substring(0,index) + "(ニコレポ非公開)";
-			}
-			infoXml = infoXml.replace("</user_id>",
-				"</user_id>\n<user>" + ownerName + "</user>");
 			try {
 				PrintWriter pw = new PrintWriter(infoFile, "UTF-8");
 				pw.write(infoXml);
@@ -1822,13 +1842,14 @@ public class ConvertWorker extends SwingWorker<String, String> {
 				else {
 					String tstr = Status.getText();
 					if(isSaveComment()) {
-						tstr = "コメント取得成功、" + tstr;
+						tstr = "コメント取得成功　" + tstr;
 					}
-					if(isSaveVideo())
-						tstr = "[警告]動画取得成功、" + tstr;
-					else
-						tstr = "[警告]" + tstr;
+					if(isSaveVideo()) {
+						tstr = "動画取得成功　" + tstr;
+					}
+					tstr = "[警告]動画情報の取得失敗　" + tstr;
 					sendtext(tstr);
+					System.out.println(tstr);
 					return result;
 				}
 			}
@@ -2606,7 +2627,7 @@ public class ConvertWorker extends SwingWorker<String, String> {
 					if(fis!=null){
 						fis.close();
 					}
-					if(fis!=null){
+					if(fos!=null){
 						fos.flush();
 						fos.close();
 					}
